Update system to (almost) give strong consistency
This is based on my sketchy conclusions in the new markdown doc in
this commit. I now write to all replicas on write, and require that to
succeed. I read from a single replica, but try the other replicas if
that fails. This makes reads fast but writes slow.

A major problem right now is that writes where only one node fails
don't rollback the change with other nodes.

There is still no data re-replication or anything to deal with data
that is under-replicated due to offline replicas. This could do with
some work – probably needs an API call to disable it when nodes are
only offline temporarily though.

There's also no vector clocks/similar. This may be an issue if writes
come in at the same time as the replicas will end up on unpredictably
different values.

Right now reads are highly available, but writes are not.