Fix unbalanced replicas
While reading about Rendezvous Hashing I'd noticed a suggested
optimisation to pre-hash the nodes and keys, then combine them with
XOR. I saw it in two different sources so implemented without
thinking about it. As ever, premature optimisation is the root of all
evil.

In that old implementation, Nodes with a high hash will have the
highest key hashes much of the time. High leading bits are preserved
through the XOR – in individual operations they can be cancelled out
by the hash, but statistically over a population they still matter.

This commit removes the XOR. Instead I compute Murmur hash for each
key and node. To keep things simple, I switched nodes to using a
32-bit hash so that it can be the seed for keys. The cluster seed
still affects it, as that's the seed for node hashes.

While looking into this I remembered hash-collision attacks exist, and
the PR that SipHash got. My system could be vulnerable to exhausting
nodes if attacked like that. Not going to switch for now though.