[WIP] Use persistent connections between nodes
gRPC seems to advise not closing connections. I'm very up for this.
But the problem is that I'm not sure how to detect 'dead' connections:
I've seen several times that a load test can break existing inter-node
gRPC connections in a way that Golang doesn't handle and I'm not sure
how to detect.