Fix requiring reads to succeed from most nodes
I only want to return reads if I've talked to most nodes. Making a
write requires talking to most nodes, so this ensures that a read
cannot miss successful writes.

This also required tweaking the examples. Until now each key-value
pair has only been stored on 2 of the 3 nodes. This doesn't give
high-availability: if 1 node goes down there's now only half the
replicas of the data it held, and that's not a majority, so reads (and
writes fail.) The examples now write data to all 3 nodes, and all data
can survive the loss of any 1 node.