Badly stop waiting for nodes to reply once quorate
Our reads and writes have to be successfully applied to a majority of
the relevant replicas. Until now, we'd wait for the 5 second timeout
on each request before carrying on.

This commit cancels requests to other replicas as soon as a majority
of nodes respond. It works… but for the data-writing queries it leaves
us with a little less data security than we'd like.

The correct, upcoming fix is to carry on without cancelling write
requests. Let optional outstanding replica writes still try to succeed
in the background, without holding up further processing.